# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Namkeen Wholesale Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products (with personalized pricing if customer is logged in)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product (admin)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
...
components:
  schemas:
...
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
      required: [ok]
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        unit:
          type: string
        basePrice:
          type: number
        customerPrice:
          type: number
          nullable: true
        imageUrl:
          type: string
          nullable: true
        inStock:
          type: boolean
        availableStock:
          type: number
        createdAt:
          type: string
      required: [id, name, unit, basePrice, inStock, availableStock, createdAt]
    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        unit:
          type: string
        basePrice:
          type: number
        imageUrl:
          type: string
          nullable: true
        inStock:
          type: boolean
      required: [name, unit, basePrice]
    UpdateProductBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        unit:
          type: string
        basePrice:
          type: number
        imageUrl:
          type: string
          nullable: true
        inStock:
          type: boolean
...